+++ static/app/views/settings/components/defaultSearchBar.tsx
